requirements.txt
streamlit==1.21.0
requests==2.31.0
beautifulsoup4==4.12.0
Pillow==10.1.0
numpy==1.25.2
scikit-learn==1.3.0
matplotlib==3.8.0